watch-price-cnn/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "watch-price-cnn"
version = "0.1.0"
description = "CNN-based watch price prediction from images (≤250K params)"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [{name = "Daniel"}]

dependencies = [
    "torch>=2.2.0",
    "torchvision>=0.17.0",
    "torchinfo>=1.8.0",
    "numpy>=1.26.0",
    "pandas>=2.1.0",
    "scikit-learn>=1.4.0",
    "matplotlib>=3.8.0",
    "seaborn>=0.13.0",
    "Pillow>=10.2.0",
    "albumentations>=1.3.1",
    "pyyaml>=6.0",
    "tqdm>=4.66.0",
    "grad-cam>=1.5.0",
]

[project.optional-dependencies]
dev = [
    "jupyter>=1.0.0",
    "ipykernel>=6.29.0",
    "black>=24.0.0",
    "ruff>=0.2.0",
    "pytest>=8.0.0",
]

[tool.setuptools.packages.find]
include = ["src*"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.black]
line-length = 100
target-version = ["py310"]
